move: RestaurantQueryRepository 이동 #1 /repository/implement -> /repository
